refactor: extract utils in validation hook
